refactor: 評価実行を結果ページ側に移動
- 「結果を見る」ボタンで即座に /result に遷移
- 結果ページで📝アニメーション付きローディング表示
- 結果ページ側でevaluate/finalEvaluateを実行
- InterviewChatから評価ロジックを完全に削除（chatのみに集中）
